+++ capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require './decorator'
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @wrapper.correct_name.capitalize
+  end
+end
