+++ nameable.rb
@@ -1,0 +1,5 @@
+class Nameable
+  def correct_name
+    raise NotImplementedError, 'correct_name method must be implemented in a subclass.'
+  end
+end
