+++ person.rb
@@ -1,16 +1,14 @@
 # Class that represents a person
-class Nameable
-  def correct_name
-    raise NotImplementedError, 'correct_name method must be implemented in a subclass.'
-  end
-end
+require './nameable'
+require './capitalize_decorator'
+require './trimmer_decorator'
 
 class Person < Nameable
   attr_accessor :name, :age
   attr_reader :id
 
   def initialize(age, name = 'Unknown', parent_permission: true)
-    super
+    super()
     @id = rand(100..1000)
     @name = name
     @age = age
@@ -31,29 +29,6 @@ class Person < Nameable
 
   def correct_name
     @name
-  end
-end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    super
-    @wrapper = nameable
-  end
-
-  def correct_name
-    @wrapper.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @wrapper.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @wrapper.correct_name[0...10]
   end
 end
